Automated Documentation

Welcome to the documentation for AirBorne!

This are the autodocs generated by Documenter.jl.

Index

AirBorne.ETL

AirBorne.ETL.YFinance

AirBorne.ETL.YFinance.get_chart_data

AirBorne.ETL.YFinance.get_interday_data

AirBorne.ETL.YFinance.hello_yfinance

AirBorne.hello_world

Autodocs

AirBorne.hello_world — Method

hello_world()

Returns a string saying "Hello World!".

AirBorne.ETL — Module

This modules centralizes all features for the data pipeline of AirBorne including API
connections, cache management and data transformations as part of the data pipeline.

AirBorne.ETL.YFinance — Module

This modules provides an interface between AirBorne and Yahoo Finance API.

AirBorne.ETL.YFinance.get_chart_data — Method

get_chart_data(symbol, period1, period2, freq)

This function calls the Yahoo chart API to get the OHCLV data. The documentation for this function is based on CryptoCoinTracker's guide.

Arguments

symbol::String: Ticker Symbol

period1::String: UNIX Timestap indicating the start of the data requested

period2::String: UNIX Timestap indicating the end time of the data requested

freq::String: The time interval between two data points. Can be 1m 2m 5m 15m 30m 60m 90m 1h 1d 5d 1wk 1mo 3mo.

Returns

r::HTTP.Messages.Response

Examples

# This example is untestable as it requires internet connection.
julia> import AirBorne
julia> r = AirBorne.ETL.YFinance.get_chart_data("AAPL","1577836800","1580515200","1d")

AirBorne.ETL.YFinance.get_interday_data — Method

function get_interday_data(symbols, period1, period2)

Use this function to get interday data for different tickers from Yahoo charts API.

Arguments

symbols::String: Ticker Symbol

period1::String: UNIX Timestap indicating the start of the data requested

period2::String: UNIX Timestap indicating the end time of the data requested

Example

import AirBorne
data = AirBorne.ETL.YFinance.get_interday_data(["AAPL","GOOG"],"1577836800","1580515200")

AirBorne.ETL.YFinance.hello_yfinance — Method

hello_yfinance()

Returns a string saying "Hello YFinance!".