Automated Documentation

Welcome to the documentation for AirBorne!

These are the autodocs generated by Documenter.jl.

Index

AirBorne.ETL

AirBorne.ETL.Cache

AirBorne.ETL.YFinance

AirBorne.Engines

AirBorne.Engines.DEDS

AirBorne.Markets

AirBorne.Markets.StaticMarket

AirBorne.Markets.StaticMarket.Order

AirBorne.Money

AirBorne.Structures.ContextTypeA

AirBorne.Wallet

AirBorne.ETL.Cache.gen_id

AirBorne.ETL.Cache.get_cache_path

AirBorne.ETL.Cache.hello_cache

AirBorne.ETL.Cache.list_bundles

AirBorne.ETL.Cache.load_bundle

AirBorne.ETL.Cache.remove_bundle

AirBorne.ETL.Cache.store_bundle

AirBorne.ETL.YFinance.get_chart_data

AirBorne.ETL.YFinance.get_interday_data

AirBorne.ETL.YFinance.hello_yfinance

AirBorne.Engines.DEDS.hello_deds

AirBorne.Engines.DEDS.run

AirBorne.Markets.StaticMarket.addJournalEntryToLedger

AirBorne.Markets.StaticMarket.addSecurityToPortfolio

AirBorne.Markets.StaticMarket.available_data

AirBorne.Markets.StaticMarket.execute_orders!

AirBorne.Markets.StaticMarket.expose_data

AirBorne.Utils.deepPush!

AirBorne.Utils.get_latest

AirBorne.Utils.hello_world

AirBorne.Utils.sortStruct!

AirBorne.Utils.sortedStructInsert!

Base.:==

Autodocs

AirBorne.Money — Type

Just a float with an associated Symbol

AirBorne.Wallet — Type

Just a wrapper around a dictionary

Base.:== — Method

Test if Wallets are identical. Potentially can be redefined as === and leave == if different keys are set to 0

AirBorne.Structures.ContextTypeA — Type

Context definitions for contract between Engines, Markets and Strategies to operate with.

AirBorne.Utils.deepPush! — Method

deepPush!(list,element)

Inserts the deepcopy of an element into a collection

AirBorne.Utils.get_latest — Method

get_latest(df,id_symbols,sort_symbol)

Retrieves last record from a dataframe, sortying by sort_symbol and grouping by id_symbols.

```julia
get_latest(past_data,[:exchangeName,:symbol],:date)
```

AirBorne.Utils.hello_world — Method

hello_world()

Returns a string saying "Hello World!".

AirBorne.Utils.sortStruct! — Method

sortStruct!(v::Vector, symbol;rev=true)

Sorts a struct by a given symbol.

AirBorne.Utils.sortedStructInsert! — Method

sortedStructInsert!(v::Vector, x,symbol;rev=true)

Inserts a struct into a sorted Vector of structs in the right place to keep the array sorted.

AirBorne.ETL — Module

This modules centralizes all features for the data pipeline of AirBorne including API
connections, cache management and data transformations as part of the data pipeline.

AirBorne.ETL.Cache — Module

This modules centralizes caching for the AirBorne package. Containing:
- Definition of data storage procedures
- Definition of data storage formats

AirBorne.ETL.Cache.gen_id — Method

gen_id()

Generates an id based on the current UTC timestamp with format "yyyy_mm_dd_H_M_S_s"

AirBorne.ETL.Cache.get_cache_path — Method

 get_cache_path()

Defines the cache path depending on the OS and environment variables.
```julia
julia> import AirBorne
julia> AirBorne.ETL.Cache.get_cache_path()
```

AirBorne.ETL.Cache.hello_cache — Method

hello_cache()

Returns a string saying "Hello Cache!".

AirBorne.ETL.Cache.list_bundles — Method

list_bundles()

Returns the list of bundles available in the cached folder.

In the future this function can be expanded to return information as timestamp, 
format of data in bundle among relevant metadata.

AirBorne.ETL.Cache.load_bundle — Method

load_bundle(bundle_id::String)

Loads data from a cached bundle.

# Returns
DataFrames.DataFrame

AirBorne.ETL.Cache.remove_bundle — Method

remove_bundle(bundle_id::String; just_archive::Bool=false)

Removes bundle from cache. This is an irreversible operation. If just_archive is true it only flushes the archive folder.

AirBorne.ETL.Cache.store_bundle — Method

store_bundle(data::DataFrames.DataFrame; bundle_id::Union{Nothing, String}=nothing, archive::Bool=true, meta::Dict=Dict(), c_meta::Dict=Dict())

Stores a dataframe in a bundle in parquet format.

**Is very important that none of the columns are of type "Any"** as the storage for this column type is not defined.

AirBorne.ETL.YFinance — Module

This modules provides an interface between AirBorne and Yahoo Finance API. 

For ForEx to obtain 5m definition the requested range must be within the last 60 days.

To request ForEx data replace the ticker by `codeA*codeB*"=X"` i.e. to obtain the ForEx of USD and GBP the 
ticker must be "USDGBP=X".

AirBorne.ETL.YFinance.get_chart_data — Method

get_chart_data(symbol, period1, period2, freq)

This function calls the Yahoo chart API to get the OHCLV data.
The documentation for this function is based on [CryptoCoinTracker's guide](https://cryptocointracker.com/yahoo-finance/yahoo-finance-api#34a6032b7b9949a1876f4568e4961afd).

# Arguments

- `symbol::String`:  Ticker Symbol
- `period1::String`:  UNIX Timestamp indicating the start of the data requested
- `period2::String`:  UNIX Timestamp indicating the end time of the data requested 
- `freq::String`: The time interval between two data points. Can be 1m 2m 5m 15m 30m 60m 90m 1h 1d 5d 1wk 1mo 3mo.

# Returns
- r::HTTP.Messages.Response

# Examples
```julia
# This example is untestable as it requires internet connection.
julia> import AirBorne
julia> r = AirBorne.ETL.YFinance.get_chart_data("AAPL","1577836800","1580515200","1d")
```

AirBorne.ETL.YFinance.get_interday_data — Method

function get_interday_data(symbols, period1, period2)

Use this function to get interday data for different tickers from Yahoo charts API.
# Arguments
- `symbols::String`:  Ticker Symbol
- `period1::String`:  UNIX Timestamp indicating the start of the data requested
- `period2::String`:  UNIX Timestamp indicating the end time of the data requested 

# Example
```julia
import AirBorne
data = AirBorne.ETL.YFinance.get_interday_data(["AAPL","GOOG"],"1577836800","1580515200")
```

AirBorne.ETL.YFinance.hello_yfinance — Method

hello_yfinance()

s a string saying "Hello YFinance!".

AirBorne.Markets — Module

Strategies

This module centralizes access to different market models.

A market model is tuple with a module and a struct that can be used to simulate/estimate the execution of orders.  

A market can represent a stock exchange like NYSE or secondary markets.

Assumptions made during the modelling of the market will have a direct impact on the results from backtesting,
any result from backtesting should be referenced to a strategy and a market model.

AirBorne.Markets.StaticMarket — Module

StaticMarket

The static market assumes that the prices of the assets are not affected by the strategy used. 

This is a big assumption for large orders but for smaller ones it can hold.

AirBorne.Markets.StaticMarket.Order — Type

Order

Represents an order to the market. I.e. A stock exchange. 

TODO: Add more documentation

AirBorne.Markets.StaticMarket.addJournalEntryToLedger — Method

addJournalEntryToLedger(ledger::Vector{Any},journalEntry::Union{DotMap,Dict})

StaticMarket method to add journal entries to ledger.

AirBorne.Markets.StaticMarket.addSecurityToPortfolio — Method

addSecurityToPortfolio(portfolio::Union{DotMap,Dict},security::Union{DotMap,Dict})

StaticMarket method to add  securities to portfolios .

AirBorne.Markets.StaticMarket.available_data — Method

available_data(context,data)

This function determine what data is available given the context.

AirBorne.Markets.StaticMarket.execute_orders! — Method

execute_orders(from, to, context,data)

This function updates the portfolio of the user that is stored in the variable context.

TODO: Add more documentation.

AirBorne.Markets.StaticMarket.expose_data — Method

expose_data(context,data)

This function determine how the data is transformed and filtered before being passed to the user.

AirBorne.Engines — Module

This modules centralizes all features for strategy backtesting including different simulation archetypes.

AirBorne.Engines.DEDS — Module

DEDS - Which stands for "Discrete Event Driven Simulation"  is a framework for backtesting
where the system moves from one event to the next one.

AirBorne.Engines.DEDS.hello_deds — Method

DEDS module hello world

AirBorne.Engines.DEDS.run — Method

run(data::DataFrame, initialize!::Function, trading_logic!::Function, execute_orders!::Function,expose_data::Function;audit=true)

Run DEDS simulation provided:
# Arguments
- `data::DataFrame`: A dataframe with the data to be provided to the function `expose_data` and `Function, execute_orders!`.
- `initialize!::Function`: initialize!(context) should receive a struct context and provide initialization for its accounts 
    and add as the next events to be processed by this function.
- `trading_logic!::Function`: `trading_logic!(context,exposed_data)` receives the context and exposed data from the market 
  and should place orders and define further events 
- `execute_orders!::Function`: execute_orders!(past_event_date, context.current_event.date, context, data) Executes orders between
    the past event and the current one, this function will modify the portfolio and the accounts from the context.
### Optional keyword arguments
- `audit::Bool=true`: If true context will its audit entry populated for each event in the simulation. 
- `max_iter::Int=10^6`: Limit the number of events processed on the simulation.
# Returns
- `context::Context`