Automated Documentation

Welcome to the documentation for AirBorne!

This are the autodocs generated by Documenter.jl.

Index

AirBorne.Backtest

AirBorne.Backtest.DEDS

AirBorne.ETL

AirBorne.ETL.Cache

AirBorne.ETL.YFinance

AirBorne.Backtest.DEDS.hello_deds

AirBorne.Backtest.DEDS.run

AirBorne.ETL.Cache.gen_id

AirBorne.ETL.Cache.get_cache_path

AirBorne.ETL.Cache.hello_cache

AirBorne.ETL.Cache.list_bundles

AirBorne.ETL.Cache.load_bundle

AirBorne.ETL.Cache.remove_bundle

AirBorne.ETL.Cache.store_bundle

AirBorne.ETL.YFinance.get_chart_data

AirBorne.ETL.YFinance.get_interday_data

AirBorne.ETL.YFinance.hello_yfinance

AirBorne.hello_world

Autodocs

AirBorne.hello_world — Method

hello_world()

Returns a string saying "Hello World!".

AirBorne.ETL — Module

This modules centralizes all features for the data pipeline of AirBorne including API
connections, cache management and data transformations as part of the data pipeline.

AirBorne.ETL.YFinance — Module

This modules provides an interface between AirBorne and Yahoo Finance API.

AirBorne.ETL.YFinance.get_chart_data — Method

get_chart_data(symbol, period1, period2, freq)

This function calls the Yahoo chart API to get the OHCLV data. The documentation for this function is based on CryptoCoinTracker's guide.

Arguments

symbol::String: Ticker Symbol

period1::String: UNIX Timestamp indicating the start of the data requested

period2::String: UNIX Timestamp indicating the end time of the data requested

freq::String: The time interval between two data points. Can be 1m 2m 5m 15m 30m 60m 90m 1h 1d 5d 1wk 1mo 3mo.

Returns

r::HTTP.Messages.Response

Examples

# This example is untestable as it requires internet connection.
julia> import AirBorne
julia> r = AirBorne.ETL.YFinance.get_chart_data("AAPL","1577836800","1580515200","1d")

AirBorne.ETL.YFinance.get_interday_data — Method

function get_interday_data(symbols, period1, period2)

Use this function to get interday data for different tickers from Yahoo charts API.

Arguments

symbols::String: Ticker Symbol

period1::String: UNIX Timestamp indicating the start of the data requested

period2::String: UNIX Timestamp indicating the end time of the data requested

Example

import AirBorne
data = AirBorne.ETL.YFinance.get_interday_data(["AAPL","GOOG"],"1577836800","1580515200")

AirBorne.ETL.YFinance.hello_yfinance — Method

hello_yfinance()

Returns a string saying "Hello YFinance!".

AirBorne.ETL.Cache — Module

This modules centralizes caching for the AirBorne package. Containing:
- Definition of data storage procedures
- Definition of data storage formats

AirBorne.ETL.Cache.gen_id — Method

gen_id()

Generates an id based on the current UTC timestamp with format "yyyy_mm_dd_H_M_S_s"

AirBorne.ETL.Cache.get_cache_path — Method

 get_cache_path()

Defines the cache path depending on the OS and environment variables.
```julia
julia> import AirBorne
julia> AirBorne.ETL.Cache.get_cache_path()
```

AirBorne.ETL.Cache.hello_cache — Method

hello_cache()

Returns a string saying "Hello Cache!".

AirBorne.ETL.Cache.list_bundles — Method

list_bundles()

Returns the list of bundles available in the cached folder.

In the future this function can be expanded to return information as timestamp, 
format of data in bundle among relevant metadata.

AirBorne.ETL.Cache.load_bundle — Method

load_bundle(bundle_id::String)

Loads data from a cached bundle.

# Returns
DataFrames.DataFrame

AirBorne.ETL.Cache.remove_bundle — Method

remove_bundle(bundle_id::String; just_archive::Bool=false)

Removes bundle from cache. This is an irreversible operation. If just_archive is true it only flushes the archive folder.

AirBorne.ETL.Cache.store_bundle — Method

store_bundle(data::DataFrames.DataFrame; bundle_id::Union{Nothing, String}=nothing, archive::Bool=true, meta::Dict=Dict(), c_meta::Dict=Dict())

Stores a dataframe in a bundle in parquet format.

**Is very important that none of the columns are of type "Any"** as the storage for this column type is not defined.

AirBorne.Backtest — Module

This modules centralizes all features for strategy backtesting including different simulation archetypes.

AirBorne.Backtest.DEDS — Module

DEDS - Which stands for "Discrete Event Driven Simulation"  is a framework for backtesting
where the system moves from one event to the next one.

AirBorne.Backtest.DEDS.hello_deds — Method

DEDS module hello world

AirBorne.Backtest.DEDS.run — Method

Run DEDS simulation